+++ packages/cubejs-clickhouse-driver/src/ClickHouseDriver.ts
@@ -1,9 +1,11 @@
 /* eslint-disable no-restricted-syntax */
 import {
   BaseDriver,
-  DownloadQueryResultsOptions, DownloadQueryResultsResult,
+  DownloadQueryResultsOptions,
+  DownloadQueryResultsResult,
   DriverInterface,
-  StreamOptions, StreamTableData, StreamTableDataWithTypes,
+  StreamOptions,
+  StreamTableDataWithTypes,
 } from '@cubejs-backend/query-orchestrator';
 import { getEnv } from '@cubejs-backend/shared';
 import genericPool, { Pool } from 'generic-pool';
@@ -19,10 +21,11 @@ const ClickhouseTypeToGeneric: Record<string, string> = {
   datetime: 'timestamp',
   datetime64: 'timestamp',
   date: 'date',
+  decimal: 'decimal',
   // integers
   int8: 'int',
   int16: 'int',
-  int32: 'bigint',
+  int32: 'int',
   int64: 'bigint',
   // unsigned int
   uint8: 'int',
@@ -176,7 +179,7 @@ export class ClickHouseDriver extends BaseDriver implements DriverInterface {
               row[field] = `${value.substring(0, 10)}T${value.substring(11, 22)}.000`;
             } else if (meta.type.includes('Date')) {
               row[field] = `${value}T00:00:00.000`;
-            } else if (meta.type.includes('Int') || meta.type.includes('Float')) {
+            } else if (meta.type.includes('Int') || meta.type.includes('Float') || meta.type.includes('Decimal')) {
               // convert all numbers into strings
               row[field] = `${value}`;
             }
@@ -206,6 +209,7 @@ export class ClickHouseDriver extends BaseDriver implements DriverInterface {
   public async stream(
     query: string,
     values: unknown[],
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     { highWaterMark }: StreamOptions
   ): Promise<StreamTableDataWithTypes> {
     // eslint-disable-next-line no-underscore-dangle